reworked onresume and onsuspend of lifecyclehandler
