+++ lib/components/map_widget.dart
@@ -4,13 +4,14 @@ import 'package:flushbar/flushbar.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:hiking4nerds/components/map_buttons.dart';
+import 'package:hiking4nerds/services/lifecycle_event_handler.dart';
 import 'package:hiking4nerds/services/sharing/geojson_data_handler.dart';
 import 'package:hiking4nerds/services/sharing/gpx_data_handler.dart';
 import 'package:hiking4nerds/services/route.dart';
 import 'package:hiking4nerds/services/routing/osmdata.dart';
 import 'package:hiking4nerds/styles.dart';
 import 'package:mapbox_gl/mapbox_gl.dart';
-import 'package:flutter/services.dart' show MethodChannel, SystemChannels, rootBundle;
+import 'package:flutter/services.dart' show MethodChannel, rootBundle;
 import 'package:location/location.dart';
 import 'package:location_permissions/location_permissions.dart';
 import 'dart:math';
@@ -53,6 +54,7 @@ class MapWidgetState extends State<MapWidget> {
   bool _zoomGesturesEnabled = true;
   bool _myLocationEnabled = true;
   bool _tilesLoaded = false;
+  bool _mapSuspended = false;
   String _currentStyle;
   Map<String, String> _styles = new Map();
   MyLocationTrackingMode _myLocationTrackingMode =
@@ -83,22 +85,9 @@ class MapWidgetState extends State<MapWidget> {
   }
 
   Future<void> _getIntentData() async {
-    SystemChannels.lifecycle.setMessageHandler((msg) {
-      if (msg.contains('resumed')) {
-        _getSharedData().then((d) {
-          if (d == null) return;
-          setState(() {
-            sharedRoute = d;
-          });
-          drawRoute(sharedRoute);
-        });
-      }
-      return;
-    });
-/*
     WidgetsBinding.instance.addObserver(
-        new LifecycleEventHandler(resumeCallBack: doCallBack)
-    );*/
+        new LifecycleEventHandler(resumeCallBack: resumeMap, suspendingCallBack: suspendMap)
+    );
 
     if (sharedRoute == null) {
       var data = await _getSharedData();
@@ -109,12 +98,20 @@ class MapWidgetState extends State<MapWidget> {
     }
   }
 
-  Future<void> doCallBack(){
+  Future<void> suspendMap(){
+    setState(() {
+      _mapSuspended = true;
+    });
+  }
+
+  Future<void> resumeMap(){
     _getSharedData().then((d) {
-      if (d == null) return;
+      if (d == null || !_mapSuspended) return;
       setState(() {
         sharedRoute = d;
+        _mapSuspended = false;
       });
+      drawRoute(sharedRoute);
     });
   }
 
@@ -547,26 +544,5 @@ class MapWidgetState extends State<MapWidget> {
     });
 
     if (sharedRoute != null) drawRoute(sharedRoute);
-  }
-}
-
-class LifecycleEventHandler extends WidgetsBindingObserver {
-  LifecycleEventHandler({this.resumeCallBack, this.suspendingCallBack});
-
-  final AsyncCallback resumeCallBack;
-  final AsyncCallback suspendingCallBack;
-
-  @override
-  Future<Null> didChangeAppLifecycleState(AppLifecycleState state) async {
-    switch (state) {
-      case AppLifecycleState.inactive:
-      case AppLifecycleState.paused:
-      case AppLifecycleState.detached:
-        await suspendingCallBack();
-        break;
-      case AppLifecycleState.resumed:
-        await resumeCallBack();
-        break;
-    }
   }
 }
